refactor: moved table components
