+++ src/client/routes/_auth.dashboard.collections.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { useAuth } from "@/client/core/providers/auth-provider";
-import { CollectionsContainer } from "../components/containers/collections-container";
+import { CollectionsContainer } from "../features/collections/containers/collections-container";
 
 export const Route = createFileRoute("/_auth/dashboard/collections")({
     component: CollectionsComponent,
